+++ braid/braid-run/src/main.rs
@@ -105,6 +105,7 @@ async fn main() -> Result<()> {
             false,
             flydra_types::StartSoftwareFrameRateLimit::Enable(cfg.framerate),
         ),
+        TriggerType::PtpSync(_) => (false, flydra_types::StartSoftwareFrameRateLimit::NoChange),
     };
     let show_tracking_params = false;